+++ GrailDiary/FileManager+DefaultContainer.swift
@@ -1,0 +1,15 @@
+//
+//  FileManager+DefaultContainer.swift
+//  GrailDiary
+//
+//  Created by Brian Dewey on 1/10/21.
+//  Copyright © 2021 Brian's Brain. All rights reserved.
+//
+
+import Foundation
+
+extension FileManager {
+  func defaultDocumentLocation(completion: @escaping (Result<URL, Error>) -> Void) {
+
+  }
+}
